+++ Product/Production/Services/AuditRepositoryCore/src/test/java/gov/hhs/fha/nhin/audit/retrieve/AuditRetrieveEventsUtilTest.java
@@ -68,7 +68,7 @@ public class AuditRetrieveEventsUtilTest {
     private final String MESSAGE_ID = "urn:uuid:4abc7cc8-1edc-409d-b061-bf65b21f7b1e";
     private final String RELATES_TO = "urn:uuid:4abc7cc8-1edc-409d-b061-bf65b21f7b1e";
     private final String USER_ID = "wanderson";
-    private final int ID = 1;
+    private final Long ID = 1L;
     private final int OUTCOME = 0;
     private final Date EVENT_TIMESTAMP = Calendar.getInstance().getTime();
     private final String AUDIT_MESSAGE = "<?xml version=\"1.0\" encoding=\"UTF-8\" standalone=\"yes\"?>"
@@ -104,7 +104,7 @@ public class AuditRetrieveEventsUtilTest {
         assertEquals("QueryAuditEventsResults.RequestMessageId mismatch", obj.getRequestMessageId(), MESSAGE_ID);
         assertEquals("QueryAuditEventsResults.RelatesTo(mismatch", obj.getRelatesTo(), RELATES_TO);
         assertEquals("QueryAuditEventsResults.UserId mismatch", obj.getUserId(), USER_ID);
-        assertEquals("QueryAuditEventsResults.id mismatch", obj.getId(), ID);
+        assertEquals("QueryAuditEventsResults.id mismatch", obj.getId(), ID.longValue());
         assertEquals("QueryAuditEventsResults.outcome mismatch", obj.getEventOutcomeIndicator().intValue(), OUTCOME);
         assertEquals("QueryAuditEventsResults.EventTimestamp mismatch",
             obj.getEventTimestamp().toGregorianCalendar().getTime(), EVENT_TIMESTAMP);
